TweetDetails added
