+++ src/views/Tweets.vue
@@ -1,15 +1,8 @@
 <template lang="pug">
   .container
     LoadingIcon(v-if="!done")
-    b-modal(id="tweetModal" title="Tweet details" ok-only)
-            p(v-html="textWithKeyword(selected.text)")
-            br
-            strong {{sentimentToText(analysis.sentiment)}} ({{Math.round(analysis.sentiment.score * 10) / 10}})
-            br
-            br
-            strong Categories:
-            ul(v-if="analysis.category instanceof Array" v-for="(category, index) in analysis.category" :key="index")
-              li {{category.name}}
+    TweetDetails(:analysis="analysis"
+              :selected="selected")
     TweetList(:tweets="tweets"
               v-if="done"
               @show-tweet="show_tweet")
@@ -18,11 +11,13 @@
 <script lang="ts">
 import { Component, Vue } from 'vue-property-decorator';
 import LoadingIcon from '@/components/LoadingIcon.vue';
+import TweetDetails from '@/components/TweetDetails.vue';
 import TweetList from '@/components/TweetList.vue';
 
 @Component({
   components: {
     LoadingIcon,
+    TweetDetails,
     TweetList
   }
 })
@@ -37,9 +32,6 @@ export default class Tweets extends Vue {
 
   // This is used to display the selected tweet in the popup when clicking on tweets
   selected = {};
-
-  // Displayed when waiting for analysis.
-  loading_text = "Loading...";
 
   // This is used to display the semantic analysis of the selected tweet.
   analysis : any = { sentiment : {}, category : {}, entities : {} };
@@ -69,10 +61,6 @@ export default class Tweets extends Vue {
   show_tweet(tweet : any) {
     this.selected = tweet;
 
-    this.analysis.sentiment = this.loading_text;
-    this.analysis.category = this.loading_text;
-    this.analysis.entities = this.loading_text;
-
     this.analyse(tweet.text, "sentiment");
     this.analyse(tweet.text, "category");
     this.analyse(tweet.text, "entities");
@@ -86,47 +74,6 @@ export default class Tweets extends Vue {
     }).catch((reason) => {
       this.analysis[type] = reason.response.data.details;
     });
-  }
-
-  textWithKeyword(text : any) {
-    if (text == undefined) return "";
-    let keyword = this.keyword(this.analysis.entities);
-    if (keyword == "") return "";
-
-    return text.replace(new RegExp(keyword.name, "gi"), (match : string) => {
-      return `<u v-b-tooltip.hover title="Importance: ${Math.round(keyword.salience * 10) / 10}" class="text-danger">${match}</u>`;
-    });
-  }
-
-  keyword(value : any) {
-    if (value instanceof Array) {
-      return JSON.parse(JSON.stringify(value)).sort((a : any, b : any) => b.salience - a.salience)[0];
-    }
-    return "";
-  }
-
-  sentimentToText(value : any) {
-    let txt = "";
-    value = value.score;
-
-    if (value <= 1 && value > 0.7) {
-        txt = "This is something VERY positive!";
-    } else if (value <= 0.7 && value > 0.3) {
-        txt = "This is positive!";
-    } else if (value <= 0.3 && value > 0) {
-        txt = "This is kinda positive.";
-    } else if (value == 0) {
-        txt = "This is neutral.";
-    } else if (value <= 0 && value > -0.3) {
-        txt = "This is kinda negative.";
-    } else if (value <= -0.3 && value > -0.7) {
-        txt = "This is negative.";
-    } else if (value <= -0.7 && value > -1) {
-        txt = "This is something VERY negative!";
-    } else {
-        txt = value
-    }
-    return txt;
   }
 
   parse_tweet(tweet : any) {
